reworked envelope, release broken

+++ casynth.c
@@ -142,9 +142,9 @@ void run_casynth( LV2_Handle handle, uint32_t nframes)
                         synth->envelope[ENV_ATTACK] = 1/(float)(*synth->env_a_p*synth->sample_rate);
                         synth->envelope[ENV_DECAY] = (*synth->env_b_p-1)/(float)(*synth->env_d_p*synth->sample_rate);
                         synth->envelope[ENV_BREAK] = *synth->env_b_p;
-                        synth->envelope[ENV_SWELL] = (*synth->env_sus_p - *synth->env_b_p)/(*synth->env_swl_p*synth->sample_rate);
+                        synth->envelope[ENV_SWELL] = (*synth->env_sus_p - *synth->env_b_p)/(float)(*synth->env_swl_p*synth->sample_rate);
                         synth->envelope[ENV_SUSTAIN] = *synth->env_sus_p;
-                        synth->envelope[ENV_RELEASE] = -*synth->env_sus_p/(*synth->env_r_p*synth->sample_rate);
+                        synth->envelope[ENV_RELEASE] = -*synth->env_sus_p/(float)(*synth->env_r_p*synth->sample_rate);
 
                         //set harmonic gains
                         if(*synth->harmonic_mode_p != synth->harmonic_mode)
